Add 62_polymorphism_benchmark: runtime vs compile-time polymorphism
- Virtual dispatch: 4315μs, indirect call via vtable
- Template dispatch: 567μs, inlined, 7.6x speedup
- Real addresses: 0x582043e599f0 (runtime), 0x582043e5a990 (compile-time)
- sizeof: BaseInsect=24, SocialInsect=32, Insect<Policy>=16
- Fixed 61_constexpr_roundoff_errors.cpp narrowing conversion

+++ src/61_constexpr_roundoff_errors.cpp
@@ -131,7 +131,7 @@ constexpr float my_fact(int n) {
         // mulss: res = res × (float)i → round-to-nearest after each mul
         // For n=15: i=15 → (float)15 = 15.0 exact → res = 87178291200 × 15 = 1307674368000
         //           1307674368000 in float → 1.307674... × 10^12 → exponent = 40 → mantissa truncated
-        res *= i;
+        res *= static_cast<float>(i);
     }
     return res;
     // RETURN for n=15: 1307674368000.0 → float approximation → 0x544B8B50 (verify with calculator)
